+++ openshift-server-mock/src/main/java/io/fabric8/openshift/client/server/mock/OpenShiftMockServerExtension.java
@@ -15,17 +15,20 @@
  */
 package io.fabric8.openshift.client.server.mock;
 
-import io.fabric8.kubernetes.client.server.mock.KubernetesCrudDispatcher;
+import io.fabric8.kubernetes.client.server.mock.KubernetesMixedDispatcher;
 import io.fabric8.kubernetes.client.server.mock.KubernetesMockServerExtension;
 import io.fabric8.mockwebserver.Context;
+import io.fabric8.mockwebserver.ServerRequest;
+import io.fabric8.mockwebserver.ServerResponse;
 import io.fabric8.openshift.client.NamespacedOpenShiftClient;
 import io.fabric8.openshift.client.OpenShiftClient;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.jupiter.api.extension.ExtensionContext;
 
 import java.lang.reflect.Field;
-import java.util.Collections;
 import java.util.HashMap;
+import java.util.Map;
+import java.util.Queue;
 
 /**
  * The class that implements JUnit5 extension mechanism. You can use it directly in your JUnit test
@@ -55,8 +58,9 @@ public class OpenShiftMockServerExtension extends KubernetesMockServerExtension 
   @Override
   protected void initializeKubernetesClientAndMockServer(Class<?> testClass) {
     EnableOpenShiftMockClient a = testClass.getAnnotation(EnableOpenShiftMockClient.class);
+    final Map<ServerRequest, Queue<ServerResponse>> responses = new HashMap<>();
     openShiftMockServer = a.crud()
-      ? new OpenShiftMockServer(new Context(), new MockWebServer(), new HashMap<>(), new KubernetesCrudDispatcher(Collections.emptyList()), a.https())
+      ? new OpenShiftMockServer(new Context(), new MockWebServer(), responses, new KubernetesMixedDispatcher(responses), a.https())
       : new OpenShiftMockServer(a.https());
     openShiftMockServer.init();
     openShiftClient = openShiftMockServer.createOpenShiftClient();